starter-template/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "my-agent"
version = "0.1.0"
description = "AgentNetwork p2p starter — replace this with your agent."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
  "anet>=1.1.0",
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "httpx>=0.25",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
include = ["my_agent*"]
